Revert "feat(semantic-tokens): add support for  DarkMode/LightMode components"
This reverts commit ac543306b71b5fca9fd014f69e378451064ad13a.
