--- .changeset/brown-moose-relate.md
@@ -1,6 +1,0 @@
----
-"@chakra-ui/color-mode": major
----
-
-Moved components `LightMode` and `DarkMode` to package `@chakra-ui/system` to
-prevent circular dependencies.
