--- .changeset/mighty-monkeys-wash.md
@@ -1,6 +1,0 @@
----
-"@chakra-ui/styled-system": minor
----
-
-Updated `_dark` and `_light` pseudo selectors to allow semantic tokens to change
-with the `DarkMode` and `LightMode` component.
