+++ .changeset/twenty-cooks-visit.md
@@ -2,6 +2,6 @@
 "@chakra-ui/system": minor
 ---
 
-Added `[data-theme]` to the CSS variables root selector. This allows the
-semantic tokens to change according to `data-theme="dark"` and
-`data-theme="light"` DOM element attributes.
+Added `[data-css-vars-root=true]` to the CSS variables root selector. This
+allows to layer the CSS variable definitions and allow the semantic tokens to
+react to `data-theme="dark"` and `data-theme="light"`.
