+++ packages/color-mode/test/color-mode-provider__server.test.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable global-require */
-import * as React from "react"
+import React from "react"
 import { render } from "@testing-library/react"
 import {
   createMockStorageManager,
