+++ packages/color-mode/test/use-color-mode-value.test.tsx
@@ -1,7 +1,11 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
-import { ColorModeProvider, useColorModeValue, useColorMode } from "../src"
+import {
+  ColorModeProvider,
+  useColorModeValue,
+  useColorMode,
+} from "../src/color-mode-provider"
 import { defaultThemeOptions } from "./utils"
 
 const lightValue = "light-value"
