+++ packages/styled-system/src/pseudos.ts
@@ -298,18 +298,12 @@ export const pseudoSelectors = {
    * Styles for when `data-theme` is applied to any parent of
    * this component or element.
    */
-  _dark:
-    ".chakra-ui-dark &:not([data-theme])," +
-    "[data-theme=dark] &:not([data-theme])," +
-    "&[data-theme=dark]",
+  _dark: ".chakra-ui-dark &, [data-theme=dark] &, &[data-theme=dark]",
   /**
    * Styles for when `data-theme` is applied to any parent of
    * this component or element.
    */
-  _light:
-    ".chakra-ui-light &:not([data-theme])," +
-    "[data-theme=light] &:not([data-theme])," +
-    "&[data-theme=light]",
+  _light: ".chakra-ui-light &, [data-theme=light] &, &[data-theme=light]",
 }
 
 export type Pseudos = typeof pseudoSelectors
