+++ packages/system/src/index.ts
@@ -1,5 +1,4 @@
 export * from "@chakra-ui/color-mode"
-export * from "./color-mode-components"
 export * from "@chakra-ui/styled-system"
 export { keyframes } from "@emotion/react"
 export type { Interpolation } from "@emotion/react"
