+++ packages/system/src/providers.tsx
@@ -34,18 +34,19 @@ export const ThemeProvider = (props: ThemeProviderProps) => {
 export interface CSSVarsProps {
   /**
    * The element to attach the CSS custom properties to.
-   * @default ":host, :root"
+   * Re-hoist CSS vars by attaching `data-css-vars-root={true}` to a DOM element.
+   * @example <div data-css-vars-root={true}></div>
+   *
+   * @default ":host, :root, [data-css-vars-root=true]"
    */
   root?: string
 }
 
-export const CSSVars = ({ root = ":host, :root" }: CSSVarsProps) => {
-  /**
-   * Append color mode selector to allow semantic tokens to change according to the color mode
-   */
-  const selector = [root, `[data-theme]`].join(",")
-  return <Global styles={(theme: any) => ({ [selector]: theme.__cssVars })} />
-}
+export const CSSVars = ({
+  root = ":host, :root, [data-css-vars-root=true]",
+}: CSSVarsProps) => (
+  <Global styles={(theme: any) => ({ [root]: theme.__cssVars })} />
+)
 
 export function useTheme<T extends object = Dict>() {
   const theme = React.useContext(
